Moved all errors into new file error.rs
