+++ src/parser.rs
@@ -1,47 +1,14 @@
 //! A family of functions which parse YAML into [`CTree`]s.
 
 use crate::{
+    error::{ParseError, TreeError},
     link::Link,
     node::Node,
-    tree::{CTree, TreeError},
+    tree::CTree,
 };
 
 use std::{fs::File, io::Read, path::Path};
 use yaml_rust::{Yaml, YamlLoader};
-
-/// A [`ParseError`] is a category of errors returned by parser functions that returns [`Result`]s.
-#[derive(Debug)]
-pub enum ParseError {
-    /// An error caused when IO issues occur during parsing.
-    IO(std::io::Error),
-    /// An error caused when YAML is unable to be scanned in.
-    Scan(yaml_rust::ScanError),
-    /// An error caused when a tree is not considered legal.
-    /// See also: [format information here](https://github.com/simbleau/convo/tree/main/examples/dialogue_files/README.md).
-    Tree(TreeError),
-    /// An error caused when validating a file for parsing.
-    Validation(String),
-}
-impl From<std::io::Error> for ParseError {
-    fn from(item: std::io::Error) -> Self {
-        ParseError::IO(item)
-    }
-}
-impl From<yaml_rust::ScanError> for ParseError {
-    fn from(item: yaml_rust::ScanError) -> Self {
-        ParseError::Scan(item)
-    }
-}
-impl From<String> for ParseError {
-    fn from(item: String) -> Self {
-        ParseError::Validation(item)
-    }
-}
-impl From<&str> for ParseError {
-    fn from(item: &str) -> Self {
-        ParseError::Validation(item.to_owned())
-    }
-}
 
 /// Try to returns a [`CTree`] which is generated from parsing a file.
 ///
